requirements.txt
tensorflow
keras
numpy
pandas
opencv-python
pillow
matplotlib
seaborn
scikit-learn
jupyter
streamlit
